Move send text functions into rTerminal
